Move endpointDisconnectRemoveEntitiesReducer
